# WebApplication1/WebApplication1/wwwroot/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NexusGear REST API",
    "version": "1.0.0",
    "description": "API đầy đủ cho nền tảng thương mại điện tử gaming gear NexusGear.\n\n**Xác thực:** Sử dụng cookie phiên (đăng nhập qua `POST /api/auth/login` trước)."
  },
  "servers": [{ "url": "" }],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"], "summary": "Danh sách sản phẩm (lọc + phân trang)",
        "parameters": [
          { "name": "search",    "in": "query", "schema": { "type": "string" } },
          { "name": "categoryId","in": "query", "schema": { "type": "integer" } },
          { "name": "brandId",   "in": "query", "schema": { "type": "integer" } },
          { "name": "minPrice",  "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice",  "in": "query", "schema": { "type": "number" } },
          { "name": "page",      "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize",  "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"], "summary": "Chi tiết sản phẩm",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Không tìm thấy" } }
      }
    },
    "/api/products/{id}/reviews": {
      "get": {
        "tags": ["Products"], "summary": "Danh sách đánh giá sản phẩm",
        "parameters": [
          { "name": "id",       "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "page",     "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["Products"], "summary": "Thêm đánh giá (cần đăng nhập)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddReviewRequest" } } } },
        "responses": { "201": { "description": "Đã gửi đánh giá" }, "409": { "description": "Đã đánh giá rồi" } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddReviewRequest": {
        "type": "object", "required": ["rating"],
        "properties": {
          "rating":  { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment": { "type": "string" }
        }
      },
...
    }
  }
}